refactor!: one SSE stream per user
Notifications and conversation messages were two endpoints, two Redis
channels and two copies of the same subscriber-and-keepalive plumbing,
delivered to the same signed-in user. They now share the channel
`events:user:<userId>` and the single endpoint `GET /conversations/stream`;
`GET /notifications/stream` is deleted along with its clients' second
connection.

Frames are unchanged, so consumers keep discriminating on `type`. The
surviving stream inherits the notification stream's acknowledged subscribe
and buffering, which the conversation stream lacked.

BREAKING CHANGE: GET /notifications/stream is removed. Subscribe to
GET /conversations/stream, which now carries notification frames too.
