+++ packages/hermes-plugin/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@indexnetwork/hermes-plugin",
-  "version": "0.34.0",
+  "version": "0.35.0",
   "description": "Hermes-native Index Network plugin with MCP and REST tools",
   "license": "MIT",
   "type": "module",
