+++ services/api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@indexnetwork/api",
-  "version": "0.112.0",
+  "version": "0.113.0",
   "license": "MIT",
   "private": true,
   "scripts": {
